+++ src/system/CategoriesModule/Controller/AdminController.php
@@ -330,6 +330,184 @@ class AdminController extends AbstractController
     }
 
     /**
+     * @Route("/update")
+     * @Method("POST")
+     *
+     * Updates a category.
+     *
+     * @param Request $request
+     *
+     * @return RedirectResponse
+     *
+     * @throws AccessDeniedException Thrown if the user doesn't have admin permissions over the module
+     */
+    public function updateAction(Request $request)
+    {
+        $this->get('zikula_core.common.csrf_token_handler')->validate($request->request->get('csrfToken'));
+
+        if (!$this->hasPermission('ZikulaCategoriesModule::', '::', ACCESS_ADMIN)) {
+            throw new AccessDeniedException();
+        }
+
+        // get data from post
+        $data = $request->request->get('category', null);
+
+        if (!isset($data['is_locked'])) {
+            $data['is_locked'] = 0;
+        }
+        if (!isset($data['is_leaf'])) {
+            $data['is_leaf'] = 0;
+        }
+        if (!isset($data['status'])) {
+            $data['status'] = 'I';
+        }
+
+        $args = [];
+
+        if ($request->request->get('category_copy', null)) {
+            $args['op'] = 'copy';
+            $args['cid'] = (int)$data['id'];
+
+            return $this->redirectToRoute('zikulacategoriesmodule_admin_op', $args);
+        }
+
+        if ($request->request->get('category_move', null)) {
+            $args['op'] = 'move';
+            $args['cid'] = (int)$data['id'];
+
+            return $this->redirectToRoute('zikulacategoriesmodule_admin_op', $args);
+        }
+
+        if ($request->request->get('category_delete', null)) {
+            $args['op'] = 'delete';
+            $args['cid'] = (int)$data['id'];
+
+            return $this->redirectToRoute('zikulacategoriesmodule_admin_op', $args);
+        }
+
+        if ($request->request->get('category_user_edit', null)) {
+            $_SESSION['category_referer'] = $request->server->get('HTTP_REFERER');
+            $args['dr'] = (int)$data['id'];
+
+            return $this->redirectToRoute('zikulacategoriesmodule_admin_edit', $args);
+        }
+
+        $processingHelper = $this->get('zikula_categories_module.category_processing_helper');
+
+        $valid = $processingHelper->validateCategoryData($data);
+        if (!$valid) {
+            $args = [
+                'mode' => 'edit',
+                'cid' => (int)$data['id']
+            ];
+
+            return $this->redirectToRoute('zikulacategoriesmodule_admin_edit', $args);
+        }
+
+        // process name
+        $data['name'] = $processingHelper->processCategoryName($data['name']);
+
+        // process parent
+        $data['parent'] = $processingHelper->processCategoryParent($data['parent_id']);
+        unset($data['parent_id']);
+
+        // process display names
+        $data['display_name'] = $processingHelper->processCategoryDisplayName($data['display_name'], $data['name']);
+
+        // get existing category
+        $entityManager = $this->get('doctrine')->getManager();
+        $category = $entityManager->find('ZikulaCategoriesModule:CategoryEntity', $data['id']);
+
+        $prevCategoryName = $category['name'];
+
+        // save category
+        $category->merge($data);
+        $entityManager->flush();
+
+        // process path and ipath
+        $category['path'] = $processingHelper->processCategoryPath($data['parent']['path'], $category['name']);
+        $category['ipath'] = $processingHelper->processCategoryIPath($data['parent']['ipath'], $category['id']);
+
+        // process category attributes
+        $attrib_names = $request->request->get('attribute_name', []);
+        $attrib_values = $request->request->get('attribute_value', []);
+        $processingHelper->processCategoryAttributes($category, $attrib_names, $attrib_values);
+
+        $entityManager->flush();
+
+        // since a name change will change the object path, we must rebuild it here
+        if ($prevCategoryName != $category['name']) {
+            $this->get('zikula_categories_module.path_builder_helper')->rebuildPaths('path', 'name', $category['id']);
+        }
+
+        $this->addFlash('status', $this->__f('Done! Saved the %s category.', ['%s' => $prevCategoryName]));
+
+        return $this->redirectToRoute('zikulacategoriesmodule_admin_view');
+    }
+
+    /**
+     * @Route("/newsave")
+     * @Method("POST")
+     *
+     * Creates a category.
+     *
+     * @param Request $request
+     *
+     * @return RedirectResponse
+     *
+     * @throws AccessDeniedException Thrown if the user doesn't have permission to add a category
+     */
+    public function newcatAction(Request $request)
+    {
+        $this->get('zikula_core.common.csrf_token_handler')->validate($request->request->get('csrfToken'));
+
+        if (!$this->hasPermission('ZikulaCategoriesModule::', '::', ACCESS_ADD)) {
+            throw new AccessDeniedException();
+        }
+
+        // get data from post
+        $data = $request->request->get('category', null);
+        $processingHelper = $this->get('zikula_categories_module.category_processing_helper');
+
+        $valid = $processingHelper->validateCategoryData($data);
+        if (!$valid) {
+            return $this->redirectToRoute('zikulacategoriesmodule_admin_newcat');
+        }
+
+        // process name
+        $data['name'] = $processingHelper->processCategoryName($data['name']);
+
+        // process parent
+        $data['parent'] = $processingHelper->processCategoryParent($data['parent_id']);
+        unset($data['parent_id']);
+
+        // process display names
+        $data['display_name'] = $processingHelper->processCategoryDisplayName($data['display_name'], $data['name']);
+
+        // save category
+        $entityManager = $this->get('doctrine')->getManager();
+        $category = new CategoryEntity();
+        $category->merge($data);
+        $entityManager->persist($category);
+        $entityManager->flush();
+
+        // process path and ipath
+        $category['path'] = $processingHelper->processCategoryPath($data['parent']['path'], $category['name']);
+        $category['ipath'] = $processingHelper->processCategoryIPath($data['parent']['ipath'], $category['id']);
+
+        // process category attributes
+        $attrib_names = $request->request->get('attribute_name', []);
+        $attrib_values = $request->request->get('attribute_value', []);
+        $processingHelper->processCategoryAttributes($category, $attrib_names, $attrib_values);
+
+        $entityManager->flush();
+
+        $this->addFlash('status', $this->__f('Done! Inserted the %s category.', ['%s' => $category['name']]));
+
+        return $this->redirectToRoute('zikulacategoriesmodule_admin_view');
+    }
+
+    /**
      * @Route("/op")
      *
      * Generic function to handle copy, delete and move operations.
